Versão inicial do controlador de acesso com projeto alterado, resolvendo erros de compilação.
