--- pacSeguranca/pom.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>br.ufg.inf.fabrica</groupId>
-        <artifactId>pac</artifactId>
-        <version>1.0-SNAPSHOT</version>
-    </parent>
-    <artifactId>pacSeguranca</artifactId>
-    <packaging>jar</packaging>
-    <properties>
-        <maven.compiler.source>1.7</maven.compiler.source>
-        <maven.compiler.target>1.7</maven.compiler.target>
-    </properties>
-    <name>pacSeguranca</name>
-    <dependencies>
-        <dependency>
-            <groupId>br.ufg.inf.fabrica</groupId>
-            <artifactId>pacDominio</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-                <groupId>org.apache.derby</groupId>
-                <artifactId>derby</artifactId>
-                <version>10.12.1.1</version>
-        </dependency>
-
-    </dependencies>
-</project>
