--- pacSeguranca/src/main/java/br/ufg/inf/fabrica/pac/seguranca/ILdapAutenticador.java
@@ -1,17 +1,0 @@
-package br.ufg.inf.fabrica.pac.seguranca;
-
-import br.ufg.inf.fabrica.pac.dominio.Usuario;
-
-
-/**
- *
- * @author danilloguimaraes
- */
-public interface ILdapAutenticador {
-
-    /**
-     * 
-     * @param usuario 
-     */
-    public Usuario autenticar(Usuario usuario);
-}
