+++ src/main/java/br/ufg/inf/fabrica/pac/seguranca/ControleAcesso.java
@@ -1,0 +1,15 @@
+package br.ufg.inf.fabrica.pac.seguranca;
+
+import br.ufg.inf.fabrica.pac.seguranca.modelo.Papel;
+import java.util.List;
+
+/**
+ *
+ * @author Danillo
+ */
+public interface ControleAcesso {
+    
+    public boolean autorizar(String recurso, List<Papel> papeis);
+    public boolean autorizaPapel(Papel papel, String recurso);
+    public boolean desautorizaPapel(Papel papel, String recurso);
+}
