+++ src/main/java/br/ufg/inf/fabrica/pac/seguranca/ILdapAutenticador.java
@@ -1,0 +1,17 @@
+package br.ufg.inf.fabrica.pac.seguranca;
+
+
+
+
+/**
+ *
+ * @author danilloguimaraes
+ */
+public interface ILdapAutenticador {
+
+    /**
+     * 
+     * @param usuario 
+     */
+    //public Usuario autenticar(Usuario usuario);
+}
