+++ src/main/java/br/ufg/inf/fabrica/pac/seguranca/dao/IPermissaoDao.java
