+++ src/main/java/br/ufg/inf/fabrica/pac/seguranca/modelo/Papel.java
